// models/home.ts
export interface HomeData {
    id: string;
    logoText: string;
    heroTitle: string;
    heroSubtitle: string;
    primaryButtonText: string;
    // primaryButtonLink: string;
    secondaryButtonText: string;
    // secondaryButtonLink: string;
    HeaderImage: {
      id: string;
      path: string; 
    }[];
  }